+++ api/signals.py
@@ -6,6 +6,7 @@ from django.dispatch import receiver
 
 from . import tasks
 from .models import ImportTask, InviteUser, UploadTask, UserProfile
+from .utils import create_objects
 
 logger = logging.getLogger(__name__)
 
@@ -59,6 +60,19 @@ def post_save_upload_task(sender, instance: UploadTask, created, **kwargs):
 
         # Start the celery task
         tasks.upload_task(instance.uuid, username, password)
+
+    if (
+        instance.status == "COMPLETED"
+        and instance.request_type == "registration"
+        and instance.data_owner
+    ):
+        create_objects(
+            instance.registration_type,
+            instance.bro_id,
+            instance.metadata,
+            instance.sourcedocument_data,
+            instance.data_owner,
+        )
 
 
 @receiver(post_save, sender=ImportTask)