+++ gmn/models.py
@@ -19,6 +19,12 @@ class GMN(models.Model):
     """
 
     uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    internal_id = models.CharField(
+        max_length=50,
+        null=True,
+        blank=True,
+        help_text="Alleen beschikbaar voor de eigenaar van de put.",
+    )
     bro_id = models.CharField(max_length=18)
     delivery_accountable_party = models.CharField(max_length=8, null=True)
     quality_regime = models.CharField(max_length=50, null=True)